调整调度平台配置

+++ scrmapp-wechat/src/main/java/com/ziwow/scrmapp/wechat/persistence/entity/EwCardActivity.java
@@ -11,7 +11,8 @@ import java.util.Date;
  */
 public class EwCardActivity {
 
-    private Integer id;
+
+    private Long id;
 
     /**
      * 是否领取（1：已领取 0：未领取）
@@ -48,11 +49,12 @@ public class EwCardActivity {
         this.type = type;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+
+    public void setId(Long id) {
         this.id = id;
     }
